Change projects numbering and update project 1 statement
