+++ projects/project0/template-project0.tex
@@ -12,13 +12,14 @@
 \usepackage{graphicx}
 \usepackage{color}
 \usepackage{hyperref}
-\usepackage[margin=2.5cm]{geometry}
+\usepackage[margin=3cm]{geometry}
 
 \begin{document}
 
+
 % ==============================================================================
 
-\title{\Large{INFO8006: Project 3 - Report}}
+\title{\Large{INFO8006: Project 1 - Report}}
 \vspace{1cm}
 \author{\small{\bf Mr Pacman - s111111} \\ \small{\bf Miss Pacman - s222222}}
 
@@ -26,30 +27,40 @@
 
 % ==============================================================================
 
-\section{Bayes filter}
+\section{Problem statement}
 
-\begin{enumerate}[label=\alph*.,leftmargin=*]
-    \item
+\begin{enumerate}[label=\alph*.,leftmargin=1.35em]
     \item
 \end{enumerate}
 
 \section{Implementation}
 
-\begin{enumerate}[label=\alph*.,leftmargin=*]
-    \item \textbf{\textit{Leave empty.}}
+\begin{enumerate}[label=\alph*.,leftmargin=1.35em]
+    \item
+    \item \textbf{{\it Leave empty.}}
+    \item
+    \item
+    \item \textbf{{\it Leave empty.}}
+    \item
 \end{enumerate}
 
-\section{Experiment}
+\section{Experiment 1}
 
-\begin{enumerate}[label=\alph*.,leftmargin=*]
+\begin{enumerate}[label=\alph*.,leftmargin=1.35em]
     \item
     \item
     \item
     \item
-    \item
-    \item
-    \item \textbf{\textit{Leave empty.}}
 \end{enumerate}
+
+\section{Experiment 2}
+
+\begin{enumerate}[label=\alph*.,leftmargin=1.35em]
+    \item
+    \item
+    \item
+\end{enumerate}
+
 
 % ==============================================================================
