+++ projects/project1/template-project1.tex
@@ -12,14 +12,14 @@
 \usepackage{graphicx}
 \usepackage{color}
 \usepackage{hyperref}
-\usepackage[margin=3cm]{geometry}
+\usepackage[margin=2.5cm]{geometry}
 
 \begin{document}
 
 
 % ==============================================================================
 
-\title{\Large{INFO8006: Project 1 - Report}}
+\title{\Large{INFO8006: Project 2 - Report}}
 \vspace{1cm}
 \author{\small{\bf Mr Pacman - s111111} \\ \small{\bf Miss Pacman - s222222}}
 
@@ -29,37 +29,34 @@
 
 \section{Problem statement}
 
-\begin{enumerate}[label=\alph*.,leftmargin=1.35em]
+\begin{enumerate}[label=\alph*.,leftmargin=*]
+    \item
+    \begin{itemize}
+        \item
+        \item
+        \item
+        \item
+        \item
+        \item
+    \end{itemize}
     \item
 \end{enumerate}
 
 \section{Implementation}
 
-\begin{enumerate}[label=\alph*.,leftmargin=1.35em]
+\begin{enumerate}[label=\alph*.,leftmargin=*]
     \item
-    \item \textbf{{\it Leave empty.}}
-    \item
-    \item
-    \item \textbf{{\it Leave empty.}}
-    \item
+    \item \textbf{\textit{Just refer to your code.}}
+    \item \textbf{\textit{Just refer to your codes, and define your three evaluation/cut-off functions pairs.}}
 \end{enumerate}
 
-\section{Experiment 1}
+\section{Experiment}
 
-\begin{enumerate}[label=\alph*.,leftmargin=1.35em]
-    \item
-    \item
+\begin{enumerate}[label=\alph*.,leftmargin=*]
     \item
     \item
 \end{enumerate}
 
-\section{Experiment 2}
-
-\begin{enumerate}[label=\alph*.,leftmargin=1.35em]
-    \item
-    \item
-    \item
-\end{enumerate}
 
 
 % ==============================================================================
