+++ projects/project2/pacmanagent.py
@@ -13,17 +13,24 @@ class PacmanAgent(Agent):
         """
         self.args = args
 
-    def get_action(self, state):
+    def get_action(self, state, belief_state):
         """
-        Given a pacman game state, returns a legal move.
+        Given a pacman game state and a belief state,
+                returns a legal move.
 
         Arguments:
         ----------
         - `state`: the current game state. See FAQ and class
                    `pacman.GameState`.
+        - `belief_state`: a list of probability matrices.
 
         Return:
         -------
         - A legal move as defined in `game.Directions`.
         """
+
+        # XXX: Your code here to obtain bonus
+
+        # XXX: End of your code here to obtain bonus
+
         return Directions.STOP
