+++ projects/project2/template-project2.tex
@@ -16,10 +16,9 @@
 
 \begin{document}
 
-
 % ==============================================================================
 
-\title{\Large{INFO8006: Project 2 - Report}}
+\title{\Large{INFO8006: Project 3 - Report}}
 \vspace{1cm}
 \author{\small{\bf Mr Pacman - s111111} \\ \small{\bf Miss Pacman - s222222}}
 
@@ -27,7 +26,7 @@
 
 % ==============================================================================
 
-\section{Problem statement}
+\section{Bayes filter}
 
 \begin{enumerate}[label=\alph*.,leftmargin=*]
     \item
@@ -37,10 +36,7 @@
 \section{Implementation}
 
 \begin{enumerate}[label=\alph*.,leftmargin=*]
-    \item
     \item \textbf{\textit{Leave empty.}}
-    \item \textbf{\textit{Leave empty.}}
-    \item
 \end{enumerate}
 
 \section{Experiment}
@@ -49,9 +45,11 @@
     \item
     \item
     \item
+    \item
+    \item
+    \item
+    \item \textbf{\textit{Leave empty.}}
 \end{enumerate}
-
-
 
 % ==============================================================================
